Added versioning in controller, services and speca
