+++ app/controllers/api/v1/standards_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Api
-  class StandardsController < BaseController
+  class V1::StandardsController < V1::BaseController
     def index
       res = Standard.pluck(:id, :standard, :section)
       records = res.each_with_object({}) {|std, result| r = {std[1] => [{name: std[2], id: std[0]}]}; result.merge!(r) {|_std, old, new| old + new }; }
