+++ config/routes.rb
@@ -12,7 +12,7 @@ Rails.application.routes.draw do
     resources :staffs, only: %i[index create]
   end
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  namespace :api, defaults: {format: :json} do
+  api_version(module: "Api::V1", path: {value: "v1"}, default: true) do
     resources :standards, only: [:index]
     resources :attendances, only: [:create] do
       collection do
