+++ spec/acceptance/sessions_spec.rb
@@ -9,7 +9,7 @@ resource "Sessions" do
   let!(:staff_1) { create(:staff_with_standards, school_id: school.id) }
   let!(:staff_2) { create(:staff_with_standards, school_id: school.id) }
 
-  get "api/staffs/sync" do
+  get "v1/staffs/sync" do
     parameter :mobile_number, "Mobile number of staff", required: true
     parameter :pin, "4 digit PIN"
     let!(:mobile_number) { staff_1.mobile_number }
