+++ spec/controllers/api/v1/base_controller_spec.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+describe Api::V1::BaseController do
+end
